Updated: CTF script
Changed: Flags and switches have a range of 130 pixels now instead of 100 to match the clients range.

Fixed: squad related commands, some of the commands weren't transferring ownership correctly because a squad owner was in the list as well and first() was called.
Arena current game stats were throwing an error when a player's turret was active but the player wasn't in the arena anymore.