Update DataPacketRcv.cs
Possible fix for "invisible laws" that are happening in USL currently, witnessed several errors yesterday "[02:47:45][Network1] DataPacketRcv line 78. System.ArgumentException: The array starting from the specified index is not long enough to read a value of the specified type. (Parameter 'value')
   at System.BitConverter.ToInt16(Byte[] value, Int32 startIndex)
   at InfServer.Protocol.DataPacketRcv.Deserialize() in /home/runner/work/Infantry-Online-Server/Infantry-Online-Server/dotnetcore/Network/Protocol/DataPacketRcv.cs:line 79 - data.Length = 1, content.Length = 3

This was the first AI recommended fix, can revert if it doesn't work.